Fixed Auth Issue & Refactored the code
