+++ client/public/index.html
@@ -24,7 +24,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>Sabka bazaar store</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
