+++ client/src/components/Carousel/carousel.module.scss
